Only include REGION configuration in Redshift COPY command if s3_bucket_region option is set
By default, the s3 bucket must be in the same region as the Redshift
cluster. If it is in a different region, then the REGION configuration
must be provided.

This change makes the s3_bucket_region option optional. It will only
need to be set if the bucket is in a different region from the Redshift
cluster.