Parse row values from MySQL result into native types
MySQL driver always returns either time.Time, nil or a byte array ([]uint8)

In the byte array case, we now parse the value into an integer if the
column type is schema.INTEGER. Otherwise, we parse it into a string.

We are planning to revisit this in the future, for now this is an MVP
approach to parsing row values.

Other drivers already parse the result into a native type.